# src/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Este documento contiene la documentación para utilizar la API de TODOs",
    "version": "1.0.0",
    "title": "TODOs API",
    "contact": {
      "email": "info@todos.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "https://ctd-todo-api.herokuapp.com",
  "basePath": "/v1",
...
  "schemes": ["https", "http"],
  "paths": {
...
    "/users": {
      "post": {
        "tags": ["users"],
        "summary": "Crea un usuario",
        "description": "Crea un nuevo usuario en la base de datos.",
        "operationId": "registerUser",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Información del usuario a crear",
            "required": true,
            "schema": {
              "$ref": "#/definitions/user_register"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operación Exitosa. Retorna un JWT.",
            "schema": {
              "type": "string",
              "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
            }
          },
          "400": {
            "description": "El usuario ya se encuentra registrado / Alguno de los datos requeridos está incompleto"
          },
          "500": {
            "description": "Error del servidor"
          }
        }
      }
    },
    "/users/login": {
      "post": {
        "tags": ["users"],
        "summary": "Loguea al usuario dentro de la API",
        "description": "",
        "operationId": "loginUser",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "Credenciales",
            "in": "body",
            "description": "El nombre de usuario y la contraseña del usuario",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operación Exitosa. Retorna un JWT.",
            "schema": {
              "type": "string",
              "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
            }
          },
          "400": {
            "description": "Contraseña incorrecta"
          },
          "404": {
            "description": "El usuario no existe"
          },
          "500": {
            "description": "Error del servidor"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "store_auth": {
      "type": "oauth2",
      "authorizationUrl": "https://ctd-todo-api.herokuapp.com/v1/users/login",
      "flow": "implicit",
      "scopes": {
        "write:tasks": "Crear y modificar tareas"
      }
    }
  },
  "definitions": {
    "user_register": {
      "type": "object",
      "required": ["firstName", "lastName", "email", "password"],
      "properties": {
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
...
  }
}